requirements.txt
asgiref==3.8.1
attrs==25.1.0
backports.zoneinfo==0.2.1
cmake==3.31.6
Django==4.2.19
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
importlib_resources==6.4.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
psycopg==3.2.5
PyJWT==2.9.0
pyproject-toml==0.0.12
pyproject-toml-api==0.1.0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.35.1
rpds-py==0.20.1
semantic-version==2.10.0
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
zipp==3.20.2
